requirements.txt
asyncpg
discord.py
redis
rich
xxhash
colorama
lru_dict
async_lru